+++ run.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 # email: khahux@163.com
 
-from app import app
+from routes import app
 
 
 if __name__ == '__main__':
